another-round

+++ administrator/components/com_scheduler/src/View/Logs/HtmlView.php
@@ -121,14 +121,15 @@ class HtmlView extends BaseHtmlView
             ->icon('icon-' . $arrow);
 
         if (\count($this->items)) {
-            $toolbar->delete('logs.delete')
-                ->message('JGLOBAL_CONFIRM_DELETE')
-                ->listCheck(true);
+            if ($canDo->get('core.delete')) {
+                $toolbar->delete('logs.delete')
+                    ->message('JGLOBAL_CONFIRM_DELETE')
+                    ->listCheck(true);
 
-
-            $toolbar->confirmButton('trash', 'COM_SCHEDULER_TOOLBAR_PURGE', 'logs.purge')
-                ->message('COM_SCHEDULER_TOOLBAR_PURGE_CONFIRM')
-                ->listCheck(false);
+                $toolbar->confirmButton('trash', 'COM_SCHEDULER_TOOLBAR_PURGE', 'logs.purge')
+                    ->message('COM_SCHEDULER_TOOLBAR_PURGE_CONFIRM')
+                    ->listCheck(false);
+            }
         }
 
         // Link to component preferences if user has admin privileges
